+++ tests/test_grains.py
@@ -12,6 +12,7 @@ from brew.constants import SI_UNITS
 from brew.exceptions import GrainException
 from brew.grains import Grain
 from brew.grains import GrainAddition
+from fixtures import BHY
 from fixtures import crystal
 from fixtures import pale
 from fixtures import pale_add
@@ -171,7 +172,7 @@ class TestGrainAdditions(unittest.TestCase):
         self.assertEqual(round(out, 2), 13.96)
         out = grain_add.get_lme_weight()
         self.assertEqual(round(out, 2), 10.47)
-        out = grain_add.get_dry_weight()
+        out = grain_add.get_dme_weight()
         self.assertEqual(round(out, 2), 8.38)
 
     def test_get_weight_lme(self):
@@ -182,7 +183,7 @@ class TestGrainAdditions(unittest.TestCase):
         self.assertEqual(round(out, 2), 13.96)
         out = grain_add.get_lme_weight()
         self.assertEqual(round(out, 2), 10.47)
-        out = grain_add.get_dry_weight()
+        out = grain_add.get_dme_weight()
         self.assertEqual(round(out, 2), 8.38)
 
     def test_get_weight_dry(self):
@@ -193,7 +194,7 @@ class TestGrainAdditions(unittest.TestCase):
         self.assertEqual(round(out, 2), 13.97)
         out = grain_add.get_lme_weight()
         self.assertEqual(round(out, 2), 10.48)
-        out = grain_add.get_dry_weight()
+        out = grain_add.get_dme_weight()
         self.assertEqual(round(out, 2), 8.38)
 
     def test_get_weight_specialty(self):
@@ -204,7 +205,7 @@ class TestGrainAdditions(unittest.TestCase):
         self.assertEqual(round(out, 2), 13.96)
         out = grain_add.get_lme_weight()
         self.assertEqual(round(out, 2), 12.42)
-        out = grain_add.get_dry_weight()
+        out = grain_add.get_dme_weight()
         self.assertEqual(round(out, 2), 9.94)
 
     def test_get_weight_map(self):
@@ -215,6 +216,42 @@ class TestGrainAdditions(unittest.TestCase):
             u'dry_weight': 8.38,
         }
         self.assertEquals(out, expected)
+
+    def test_convert_to_cereal(self):
+        grain_add = GrainAddition(pale,
+                                  weight=1.0,
+                                  grain_type=GRAIN_TYPE_CEREAL)
+        ga_cereal_weight = grain_add.convert_to_cereal(brew_house_yield=BHY).weight  # noqa
+        ga_lme_weight = grain_add.convert_to_lme(brew_house_yield=BHY).weight
+        ga_dme_weight = grain_add.convert_to_dme(brew_house_yield=BHY).weight
+
+        self.assertEquals(grain_add.weight, ga_cereal_weight)
+        self.assertEquals(ga_lme_weight, 0.75 * BHY)
+        self.assertEquals(ga_dme_weight, 0.6 * BHY)
+
+    def test_convert_to_lme(self):
+        grain_add = GrainAddition(pale,
+                                  weight=1.0,
+                                  grain_type=GRAIN_TYPE_LME)
+        ga_cereal_weight = grain_add.convert_to_cereal(brew_house_yield=BHY).weight  # noqa
+        ga_lme_weight = grain_add.convert_to_lme(brew_house_yield=BHY).weight
+        ga_dme_weight = grain_add.convert_to_dme(brew_house_yield=BHY).weight
+
+        self.assertEquals(round(ga_cereal_weight, 2), 1.90)
+        self.assertEquals(grain_add.weight, ga_lme_weight)
+        self.assertEquals(ga_dme_weight, 0.8)
+
+    def test_convert_to_dme(self):
+        grain_add = GrainAddition(pale,
+                                  weight=1.0,
+                                  grain_type=GRAIN_TYPE_DME)
+        ga_cereal_weight = grain_add.convert_to_cereal(brew_house_yield=BHY).weight  # noqa
+        ga_lme_weight = grain_add.convert_to_lme(brew_house_yield=BHY).weight
+        ga_dme_weight = grain_add.convert_to_dme(brew_house_yield=BHY).weight
+
+        self.assertEquals(round(ga_cereal_weight, 2), 2.38)
+        self.assertEquals(ga_lme_weight, 1.25)
+        self.assertEquals(grain_add.weight, ga_dme_weight)
 
     def test_eq(self):
         grain_add1 = GrainAddition(pale, weight=13.96)