+++ tests/test_recipes.py
@@ -46,7 +46,7 @@ class TestRecipe(unittest.TestCase):
                         grain_additions=grain_additions,
                         hop_additions=hop_additions,
                         yeast=yeast,
-                        percent_brew_house_yield=0.70,
+                        brew_house_yield=0.70,
                         start_volume=7.0,
                         final_volume=5.0,
                         )
@@ -58,7 +58,7 @@ class TestRecipe(unittest.TestCase):
 
     def test_repr(self):
         out = repr(self.recipe)
-        self.assertEquals(out, u"Recipe('pale ale', grain_additions=[GrainAddition(Grain('pale 2-row', color=2.0, hwe=308.78), weight=13.96, grain_type='cereal', units='imperial'), GrainAddition(Grain('crystal C20', color=20.0, hwe=292.09), weight=0.78, grain_type='cereal', units='imperial')], hop_additions=[HopAddition(Hop('centennial', percent_alpha_acids=0.14), weight=0.57, boil_time=60.0, hop_type='pellet', utilization_cls=HopsUtilizationGlennTinseth, units='imperial'), HopAddition(Hop('cascade', percent_alpha_acids=0.07), weight=0.76, boil_time=5.0, hop_type='pellet', utilization_cls=HopsUtilizationGlennTinseth, units='imperial')], yeast=Yeast('Wyeast 1056', percent_attenuation=0.75), percent_brew_house_yield=0.7, start_volume=7.0, final_volume=5.0, units=imperial)")  # noqa
+        self.assertEquals(out, u"Recipe('pale ale', grain_additions=[GrainAddition(Grain('pale 2-row', color=2.0, hwe=308.78), weight=13.96, grain_type='cereal', units='imperial'), GrainAddition(Grain('crystal C20', color=20.0, hwe=292.09), weight=0.78, grain_type='cereal', units='imperial')], hop_additions=[HopAddition(Hop('centennial', percent_alpha_acids=0.14), weight=0.57, boil_time=60.0, hop_type='pellet', utilization_cls=HopsUtilizationGlennTinseth, units='imperial'), HopAddition(Hop('cascade', percent_alpha_acids=0.07), weight=0.76, boil_time=5.0, hop_type='pellet', utilization_cls=HopsUtilizationGlennTinseth, units='imperial')], yeast=Yeast('Wyeast 1056', percent_attenuation=0.75), brew_house_yield=0.7, start_volume=7.0, final_volume=5.0, units=imperial)")  # noqa
 
     def test_eq(self):
         recipe1 = Recipe(u'pale ale')
@@ -91,11 +91,11 @@ class TestRecipe(unittest.TestCase):
                          yeast=None)
         self.assertTrue(recipe1 != recipe2)
 
-    def test_ne_percent_brew_house_yield(self):
+    def test_ne_brew_house_yield(self):
         recipe1 = Recipe(u'pale ale',
-                         percent_brew_house_yield=0.70)
+                         brew_house_yield=0.70)
         recipe2 = Recipe(u'pale ale',
-                         percent_brew_house_yield=0.65)
+                         brew_house_yield=0.65)
         self.assertTrue(recipe1 != recipe2)
 
     def test_ne_start_volume(self):
@@ -182,7 +182,7 @@ class TestRecipeBuilder(unittest.TestCase):
 
     def test_repr(self):
         out = repr(self.builder)
-        self.assertEquals(out, u"RecipeBuilder('pale ale', grain_list=[Grain('pale 2-row', color=2.0, hwe=308.78), Grain('crystal C20', color=20.0, hwe=292.09)], hop_list=[Hop('centennial', percent_alpha_acids=0.14), Hop('cascade', percent_alpha_acids=0.07)], target_og=1.0761348, percent_brew_house_yield=0.7, start_volume=7.0, final_volume=5.0, units=imperial)")  # noqa
+        self.assertEquals(out, u"RecipeBuilder('pale ale', grain_list=[Grain('pale 2-row', color=2.0, hwe=308.78), Grain('crystal C20', color=20.0, hwe=292.09)], hop_list=[Hop('centennial', percent_alpha_acids=0.14), Hop('cascade', percent_alpha_acids=0.07)], target_og=1.0761348, brew_house_yield=0.7, start_volume=7.0, final_volume=5.0, units=imperial)")  # noqa
 
     def test_eq(self):
         builder1 = RecipeBuilder(u'pale ale')
@@ -208,11 +208,11 @@ class TestRecipeBuilder(unittest.TestCase):
                                  hop_list=[hop_list[0]])
         self.assertTrue(builder1 != builder2)
 
-    def test_ne_percent_brew_house_yield(self):
+    def test_ne_brew_house_yield(self):
         builder1 = RecipeBuilder(u'pale ale',
-                                 percent_brew_house_yield=0.70)
+                                 brew_house_yield=0.70)
         builder2 = RecipeBuilder(u'pale ale',
-                                 percent_brew_house_yield=0.65)
+                                 brew_house_yield=0.65)
         self.assertTrue(builder1 != builder2)
 
     def test_ne_start_volume(self):